seprate test from source; fix Pt::IsSize
