+++ Makefile
@@ -1,7 +1,9 @@
 NAME := _stitch
 SRC_DIR := src
+TEST_DIR := test
 SRC = $(wildcard $(SRC_DIR)/*.cpp)
 INC = $(wildcard $(SRC_DIR)/*.hpp)
+TEST = $(wildcard $(TEST_DIR)/*.cpp)
 
 CXX := g++
 CXX_FLAGS := -std=c++17 -Wall -fPIC
@@ -18,8 +20,8 @@ py: $(SRC) $(INC)
 	@pip install pybind11
 	$(CXX) $(CXX_FLAGS) $(PY_FLAGS) $(SRC) -o $(NAME).so
 
-gtest: $(SRC) $(INC)
-	$(CXX) $(CXX_FLAGS) $(GTEST_FLAGS) $(SRC) -o $(NAME)
+gtest: $(SRC) $(INC) $(TEST)
+	$(CXX) $(CXX_FLAGS) $(GTEST_FLAGS) $(SRC) $(TEST) -o $(NAME)
 	./$(NAME)
 
 .PHONY: clean
