+++ test/gtest.cpp
@@ -1,8 +1,5 @@
 #include <gtest/gtest.h>
 
-#include "stitch.hpp"
-#include "tile.hpp"
-
 int main(int argc, char **argv) {
   ::testing::InitGoogleTest(&argc, argv);
   return RUN_ALL_TESTS();
